requirements.txt
langchain==0.1.4
pandas==2.0.1
tiktoken==0.4.0
openai==0.27.2
langchain_google_genai==0.0.4
gradio==3.50.2
datasets==2.15.0
tiktoken==0.4.0
func_timeout==4.3.5
python-dotenv